refactor(ratio-ui): move next components to separate lib
