+++ apps/web/src/app/admin/organizations/page.tsx
@@ -6,7 +6,7 @@ import Wrapper from '@/components/eventuras/Wrapper';
 import withAuthorization from '@/utils/auth/withAuthorization';
 import { createClient } from '@/utils/apiClient';
 import {List} from '@eventuras/ratio-ui/core/List';
-import { Link } from '@eventuras/ratio-ui/next/Link';
+import { Link } from '@eventuras/ratio-ui-next/Link';
 
 const AdminOrganizationsPage = async () => {
   const t = await getTranslations();
