+++ apps/web/src/app/admin/users/UserList.tsx
@@ -3,7 +3,7 @@ import { createColumnHelper, DataTable } from '@eventuras/datatable';
 import { UserDto } from '@eventuras/sdk';
 import { useTranslations } from 'next-intl';
 
-import { Link } from '@eventuras/ratio-ui/next/Link';
+import { Link } from '@eventuras/ratio-ui-next/Link';
 const columnHelper = createColumnHelper<UserDto>();
 interface UserListProps {
   users: UserDto[];
