+++ apps/web/src/app/events/EventNotFound.tsx
@@ -3,7 +3,7 @@
 import { Heading, Text } from '@eventuras/ratio-ui';
 import { useTranslations } from 'next-intl';
 
-import { Link } from '@eventuras/ratio-ui/next/Link';
+import { Link } from '@eventuras/ratio-ui-next/Link';
 
 export default function EventNotFound() {
   const t = useTranslations();
