+++ apps/web/src/components/user/UserProfileCard.tsx
@@ -6,7 +6,7 @@ import { useTranslations } from 'next-intl';
 
 import { Card } from '@eventuras/ratio-ui/core/Card';
 import {Text} from '@eventuras/ratio-ui/core/Text';
-import { Link } from '@eventuras/ratio-ui/next/Link';
+import { Link } from '@eventuras/ratio-ui-next/Link';
 
 export type UserProfileCardProps = {
   profile: UserDto;
