+++ libs/ratio-ui-next/package.json
@@ -1,0 +1,43 @@
+{
+  "name": "@eventuras/ratio-ui-next",
+  "version": "0.1.0",
+  "type": "module",
+  "exports": {
+    "./Image": {
+      "import": "./dist/Image/index.js",
+      "types": "./dist/Image/index.d.ts"
+    },
+    "./Link": {
+      "import": "./dist/Link/index.js",
+      "types": "./dist/Link/index.d.ts"
+    }
+  },
+  "main": "./dist/index.js",
+  "module": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsc && vite build"
+  },
+  "dependencies": {
+    "@eventuras/ratio-ui": "workspace:*"
+  },
+  "devDependencies": {
+    "@eventuras/eslint-config": "workspace:*",
+    "@eventuras/typescript-config": "workspace:*",
+    "@types/react": "^19.2.2",
+    "@types/react-dom": "^19.2.2",
+    "@vitejs/plugin-react": "^5.0.4",
+    "next": "^15.5.5",
+    "typescript": "5.9.3",
+    "vite": "7.1.10",
+    "vite-plugin-dts": "4.5.4"
+  },
+  "peerDependencies": {
+    "next": "^15.0.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0"
+  }
+}
