+++ libs/ratio-ui-next/src/index.ts
@@ -1,0 +1,2 @@
+export { Image } from './Image';
+export { Link } from './Link';
